+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+
+### Added
+- #690 Submissions in assignment view show how many review comments they have
+- #609 Users will be notified if they receive a comment on their submission
+
+## [1.5.5] - 2021-10-22
+
 ### Added
 - #589 When a teacher adds a new review comment about a student's submission, an eye symbol on the same submission will turn 
   red on the student's view. When submission is opened by the student, the red color will be removed again.